Uprev minijail and refactor Makefile
This change refactors the Makefile to make it easier to build omegajail
with different flags (e.g. to enable ASan or to make it link
statically). That also fixed some inconsistencies in the build, like not
adding `-Wall -Werror` to all files or adding `-fexceptions` to files
that didn't need it.