Make cgroups-related files read-only
After we've set the memory limit in bytes, we don't want the sandboxed
process to be able to raise it. In order to do so, we chmod(2) the file
so that the process cannot write to it. We also rely on the fact that
none of the seccomp-bpf scripts allow chmod(2) to undo the restriction.