Refactor how the Docker images are built  #minor
It turns out that since Docker images are additive, and the runtime
container was depending on the container that runs mkroot, and that
generates _tons_ of files, the resulting container ended up having ~2x
the number of files, just to be deleted immediately in the runtime
container. This caused zero disk savings.

Now the container that runs mkroot is a completely separate one, so
there's no duplication in the runtime container!