Use Linux 5.4's seccomp-bpf userspace notifications
This change finally gets rid of all the weird SIGSYS detectors in favor
of seccomp-bpf userspace notifications. Now all the forbidden syscalls
are being tracked, regardless of which process triggered it, since
before only the first process in the tree would get the correct SIGSYS
notification to the pid1 process.

This also now works with a debugger and `strace(1)`, since `ptrace(2)`
is now officially supported!